+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Filter navigation aids by prefix and between two fixes
+
 ## [0.7.1] - 2026-04-22
 
 ### Fixed
@@ -159,6 +165,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Aircraft performance entry.
 - Fuel planning and Mass & Balance.
 
+[unreleased]: https://github.com/AeronauticalMaps/libefb/compare/v0.7.1...HEAD
 [0.7.1]: https://github.com/AeronauticalMaps/libefb/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/AeronauticalMaps/libefb/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/AeronauticalMaps/libefb/compare/v0.5.0...v0.6.0